Fixes ported up from BANSHEE_0_10_9_SL:
2006-05-22  Aaron Bockover  <aaron@abock.org>

    Fixes ported up from BANSHEE_0_10_9_SL:

    * src/Banshee.Base/PlayerEngineCore.cs (set_Volume): pass on value
    to every loaded engine, not just active engine (BNC #176255)

    * src/Main.cs: Call dbus_g_thread_init after Gtk.Application.Init();
    fixes the CD burning crash seen on SMP machines (BNC #175378)

    * src/PlaylistView.cs: Use a custom IComparer to sort columns by desired
    order before attaching to view; use AppendColumn on sorted columns instead
    of InsertColumn (BNC #120666)

    * src/Banshee.Base/Database.cs: Added ViewOrder column to PlaylistEntries
    table schema; check to see if ViewOrder exists on an existing
    PlaylistEntries table and if not, alter the table accordingly (BNC #168653)

    * src/Banshee.Base/Sources/PlaylistSource.cs: No longer sync local
    playlist back to database after every minor change; operations are now
    directly translated into SQL queries (append, delete, reorder) to bring
    great speed improvements when handling large playlists (BNC #168653)

    * src/PlayerInterface.cs: Reset vertical scroll adjustment of the
    playlist view to the upper bound minus the page size if the value is
    greater than that when changing sources (BNC #174406)

+++ src/Banshee.Base/Sources/PlaylistSource.cs
@@ -1,9 +1,8 @@
-
 /***************************************************************************
  *  PlaylistSource.cs
  *
- *  Copyright (C) 2005 Novell
- *  Written by Aaron Bockover (aaron@aaronbock.net)
+ *  Copyright (C) 2005-2006 Novell, Inc.
+ *  Written by Aaron Bockover <aaron@abock.org>
  ****************************************************************************/
 
 /*  THIS FILE IS LICENSED UNDER THE MIT LICENSE AS OUTLINED IMMEDIATELY BELOW: 
@@ -52,6 +51,8 @@ namespace Banshee.Sources
         }
         
         private ArrayList tracks = new ArrayList();
+        private Queue remove_queue = new Queue();
+        private Queue append_queue = new Queue();
         private int id;
 
         public int Id {
@@ -88,18 +89,39 @@ namespace Banshee.Sources
         }
         
         private void LoadFromDatabase()
-        {
+        {   
             Name = (string)Globals.Library.Db.QuerySingle(String.Format(
                 @"SELECT Name
                     FROM Playlists
                     WHERE PlaylistID = '{0}'",
                     id
             ));
-            
+         
+            // check to see if ViewOrder has ever been set, if not, perform
+            // a default ordering as a compatibility update
+            if(Convert.ToInt32(Globals.Library.Db.QuerySingle(String.Format(
+                @"SELECT COUNT(*) 
+                    FROM PlaylistEntries
+                    WHERE PlaylistID = '{0}'
+                        AND ViewOrder > 0",
+                    id))) <= 0) {
+                Console.WriteLine("Performing compatibility update on playlist '{0}'", Name);
+                Globals.Library.Db.Execute(String.Format(
+                    @"UPDATE PlaylistEntries
+                        SET ViewOrder = (ROWID -
+                            (SELECT COUNT(*) 
+                                FROM PlaylistEntries
+                                WHERE PlaylistID < '{0}'))
+                        WHERE PlaylistID = '{0}'",
+                        id
+                ));
+            }
+   
             IDataReader reader = Globals.Library.Db.Query(String.Format(
                 @"SELECT TrackID 
                     FROM PlaylistEntries
-                    WHERE PlaylistID = '{0}'",
+                    WHERE PlaylistID = '{0}'
+                    ORDER BY ViewOrder",
                     id
             ));
             
@@ -147,6 +169,7 @@ namespace Banshee.Sources
             if(track is LibraryTrackInfo) {
                 lock(TracksMutex) {
                     tracks.Add(track);
+                    append_queue.Enqueue(track);
                 }
                 OnUpdated();
             }
@@ -161,6 +184,7 @@ namespace Banshee.Sources
         {
             lock(TracksMutex) {
                 tracks.Remove(track);
+                remove_queue.Enqueue(track);
             }
         }
 
@@ -183,36 +207,88 @@ namespace Banshee.Sources
                     id
             ));
             
+            tracks.Clear();
+            append_queue.Clear();
+            remove_queue.Clear();
+            
             SourceManager.RemoveSource(this);
             playlists.Remove(this);
         }
         
         public override void Commit()
         {
-            Globals.Library.Db.Execute(String.Format(
-                @"DELETE FROM PlaylistEntries
-                    WHERE PlaylistID = '{0}'",
-                    id
-            ));
+            if(remove_queue.Count > 0) {
+                lock(TracksMutex) {
+                    while(remove_queue.Count > 0) {
+                        TrackInfo track = remove_queue.Dequeue() as TrackInfo;
+                        Globals.Library.Db.Execute(String.Format(
+                            @"DELETE FROM PlaylistEntries
+                                WHERE PlaylistID = '{0}'
+                                AND TrackID = '{1}'",
+                                id, track.TrackId
+                        ));
+                    }
+                }
+            }
             
-            lock(TracksMutex) {
-                foreach(TrackInfo track in Tracks) {
-                    if(track.TrackId <= 0)
-                        continue;
-                        
-                    Globals.Library.Db.Execute(String.Format(
-                        @"INSERT INTO PlaylistEntries 
-                            VALUES (NULL, '{0}', '{1}')",
-                            id, track.TrackId
-                    ));
+            if(append_queue.Count > 0) {
+                lock(TracksMutex) {
+                    while(append_queue.Count > 0) {
+                        TrackInfo track = append_queue.Dequeue() as TrackInfo;
+                        Globals.Library.Db.Execute(String.Format(
+                            @"INSERT INTO PlaylistEntries 
+                                VALUES (NULL, '{0}', '{1}', (
+                                    SELECT CASE WHEN MAX(ViewOrder)
+                                        THEN MAX(ViewOrder) + 1
+                                        ELSE 1 END
+                                    FROM PlaylistEntries 
+                                    WHERE PlaylistID = '{0}')
+                                )", id, track.TrackId
+                        ));
+                    }
                 }
             }
         }
         
         public override void Reorder(TrackInfo track, int position)
         {
-            RemoveTrack(track);
             lock(TracksMutex) {
+                int sql_position = 1;
+            
+                if(position > 0) {
+                    TrackInfo sibling = tracks[position] as TrackInfo;
+                    if(sibling == track || sibling == null) {
+                        return;
+                    }
+                    
+                    sql_position = Convert.ToInt32(Globals.Library.Db.QuerySingle(String.Format(
+                        @"SELECT ViewOrder
+                            FROM PlaylistEntries
+                            WHERE PlaylistID = '{0}'
+                                AND TrackID = '{1}'
+                            LIMIT 1", id, sibling.TrackId)
+                    ));
+                } else if(tracks[position] == track) {
+                    return;
+                } 
+                
+                Globals.Library.Db.Execute(String.Format(
+                    @"UPDATE PlaylistEntries
+                        SET ViewOrder = ViewOrder + 1
+                        WHERE PlaylistID = '{0}'
+                            AND ViewOrder >= '{1}'",
+                    id, sql_position
+                ));
+                
+                Globals.Library.Db.Execute(String.Format(
+                    @"UPDATE PlaylistEntries
+                        SET ViewOrder = '{1}'
+                        WHERE PlaylistID = '{0}'
+                            AND TrackID = '{2}'",
+                    id, sql_position, track.TrackId
+                ));
+                
+                tracks.Remove(track);
                 tracks.Insert(position, track);
             }
         }
@@ -237,10 +313,13 @@ namespace Banshee.Sources
             
             int removed_count = 0;
             
-            foreach(TrackInfo track in args.Tracks) {
-                if(tracks.Contains(track)) {
-                    tracks.Remove(track);
-                    removed_count++;
+            lock(TracksMutex) {
+                foreach(TrackInfo track in args.Tracks) {
+                    if(tracks.Contains(track)) {
+                        tracks.Remove(track);
+                        remove_queue.Enqueue(track);
+                        removed_count++;
+                    }
                 }
             }
